+++ src/dataconnect/ensureApis.ts
@@ -3,10 +3,10 @@ import { ensure } from "../ensureApiEnabled";
 
 const prefix = "dataconnect";
 
-export async function ensureApis(projectId: string): Promise<void> {
+export async function ensureApis(projectId: string, silent: boolean = false): Promise<void> {
   await Promise.all([
-    ensure(projectId, api.dataconnectOrigin(), prefix),
-    ensure(projectId, api.cloudSQLAdminOrigin(), prefix),
+    ensure(projectId, api.dataconnectOrigin(), prefix, silent),
+    ensure(projectId, api.cloudSQLAdminOrigin(), prefix, silent),
   ]);
 }