+++ src/firestore/api-types.ts
@@ -183,6 +183,36 @@ export interface DatabaseResp {
   databaseEdition?: DatabaseEdition;
 }
 
+export interface BulkDeleteDocumentsRequest {
+  // Database to operate. Should be of the form:
+  // `projects/{project_id}/databases/{database_id}`.
+  name: string;
+  // IDs of the collection groups to delete. Unspecified means *all* collection groups.
+  // Each collection group in this list must be unique.
+  collectionIds?: string[];
+}
+
+export type BulkDeleteDocumentsResponse = {
+  name?: string;
+};
+
+export interface Operation {
+  name: string;
+  done: boolean;
+  metadata: Record<string, any>;
+  response?: Record<string, any>;
+  error?: {
+    name: string;
+    message: string;
+    code: number;
+    details?: any[];
+  };
+}
+
+export interface ListOperationsResponse {
+  operations: Operation[];
+}
+
 export interface RestoreDatabaseReq {
   databaseId: string;
   backup: string;